feat : error 형식 수정
