+++ src/main/java/com/prgrms/monthsub/common/error/exception/domain/article/ArticleException.java
@@ -1,0 +1,5 @@
+package com.prgrms.monthsub.common.error.exception.domain.article;
+
+public class ArticleException {
+
+}
