+++ src/main/java/com/prgrms/monthsub/common/error/exception/domain/series/SeriesException.java
@@ -1,0 +1,17 @@
+package com.prgrms.monthsub.common.error.exception.domain.series;
+
+import com.prgrms.monthsub.common.error.ErrorCodes;
+import com.prgrms.monthsub.common.error.exception.global.BusinessException;
+import java.util.Arrays;
+
+public class SeriesException {
+
+    public static class SeriesNotFound extends BusinessException {
+
+        public SeriesNotFound(String... message) {
+            super(ErrorCodes.ENTITY_NOT_FOUND(Arrays.stream(message).toList().toString()));
+        }
+
+    }
+
+}
