+++ src/main/java/com/prgrms/monthsub/common/error/exception/domain/writer/WriterException.java
@@ -1,0 +1,5 @@
+package com.prgrms.monthsub.common.error.exception.domain.writer;
+
+public class WriterException {
+
+}
