+++ src/main/java/com/prgrms/monthsub/common/error/exception/global/S3Exception.java
@@ -1,0 +1,5 @@
+package com.prgrms.monthsub.common.error.exception.global;
+
+public class S3Exception {
+
+}
