--- src/main/java/com/prgrms/monthsub/common/error/exception/user/AuthenticationException.java
@@ -1,5 +1,0 @@
-package com.prgrms.monthsub.common.error.exception.user;
-
-public class AuthenticationException {
-
-}
