--- src/main/java/com/prgrms/monthsub/common/error/exception/user/S3Exception.java
@@ -1,5 +1,0 @@
-package com.prgrms.monthsub.common.error.exception.user;
-
-public class S3Exception {
-
-}
