+++ src/main/java/com/prgrms/monthsub/service/AuthenticationService.java
@@ -1,6 +1,6 @@
 package com.prgrms.monthsub.service;
 
-import com.prgrms.monthsub.common.error.exception.user.UserException.UserNotFound;
+import com.prgrms.monthsub.common.error.exception.global.AuthenticationException.UserNotExist;
 import com.prgrms.monthsub.domain.User;
 import com.prgrms.monthsub.repository.UserRepository;
 import org.springframework.stereotype.Service;
@@ -18,7 +18,7 @@ public class AuthenticationService {
 
     public User findByUserName(String username) {
         return userRepository.findByUsername(username)
-            .orElseThrow(() -> new UserNotFound("username=" + username));
+            .orElseThrow(() -> new UserNotExist("username=" + username));
     }
 
 }
