+++ src/main/java/com/prgrms/monthsub/service/S3UploaderService.java
@@ -10,7 +10,7 @@ import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.amazonaws.util.IOUtils;
 import com.prgrms.monthsub.common.error.ErrorCodes;
-import com.prgrms.monthsub.common.error.exception.BusinessException;
+import com.prgrms.monthsub.common.error.exception.global.BusinessException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Arrays;
